requirement.txt
appdirs==1.4.3
asgiref==3.2.7
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.14
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-embed-video==1.3.3
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.2.0
mutagen==1.45.0
nepali==0.4.0
nepali-date==2.0.5
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.3.0.36
packaging==20.3
phonenumbers==8.12.6
Pillow==7.1.1
pipenv==2020.6.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyparsing==2.4.6
pytest==5.4.1
python3-openid==3.1.0
pytz==2018.9
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.15
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
Werkzeug==1.0.1
xxx==0.0.1
